further code optimizations
