+++ database/populate.py
@@ -1,4 +1,3 @@
-from database import users
 from models.user import User
 from settings import db
 from models.painter import Painter
@@ -6,7 +5,6 @@ from models.painting import Painting
 
 #creates tables (schema)
 db.create_all()
-db.session.commit()
 
 painters = {
     'painter1': Painter('Van Gogh1'),
@@ -30,6 +28,6 @@ for painter in painters.values():
 for painting in paintings:
     db.session.add(painting)
 
-users.add_user('test', 'test')
+User.add_user('test', 'test')
 
 db.session.commit()
