+++ settings.py
@@ -11,3 +11,4 @@ SECRET_KEY = 'development key'
 app = Flask(__name__)
 app.config.from_object(__name__)
 db = SQLAlchemy(app)
+salt = 'asdfas213540808ljl(df'
